requirements.txt
genesis_world
geoopt==0.5.1
hydra-core==1.3.2
torch-geometric==2.6.1
h5py==3.14.0
tqdm==4.67.1
scikit-learn